Fixed removing steam-dir games via the toolstrip.

+++ src/TerrariaDepotDownloader/MainForm.cs
@@ -812,19 +812,6 @@ namespace TerrariaDepotDownloader
         // Remove App Tool Via ToolStrip
         private void ToolStripMenuItem3_Click(object sender, EventArgs e)
         {
-            // Check If Removal Is Avalible
-            if (!button5.Enabled)
-            {
-                return;
-            }
-
-            // Disable If Use Steam Directory Enabled
-            if (checkBox2.Checked)
-            {
-                MessageBox.Show("You cannot use this feature while \"Use Steam Directory\" feature is enabled.", "TerrariaDepotDownloader v" + FileVersionInfo.GetVersionInfo(Path.GetFileName(System.Windows.Forms.Application.ExecutablePath)).FileVersion, MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
-                return;
-            }
-
             // Get Each Row
             foreach (ListViewItem itemRow in this.listView1.Items)
             {
@@ -835,29 +822,67 @@ namespace TerrariaDepotDownloader
                     if (itemRow.SubItems[2].Text == "Yes")
                     {
                         // Check If Client Is Currently Running - Update 1.8.3
-                        bool isRunning = Process.GetProcessesByName("Terraria").FirstOrDefault(p => p.MainModule.FileName.StartsWith(Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text, StringComparison.InvariantCultureIgnoreCase)) != default(Process);
-                        if (isRunning)
+                        // Check if use steam directory.
+                        if (!checkBox2.Checked)
                         {
-                            // Is running
-                            foreach (var process in Process.GetProcessesByName("Terraria"))
+                            bool isRunning = Process.GetProcessesByName("Terraria").FirstOrDefault(p => p.MainModule.FileName.StartsWith(Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text, StringComparison.InvariantCultureIgnoreCase)) != default(Process);
+                            if (isRunning)
                             {
-                                process.Kill();
-
-                                // Log Item
-                                if (checkBox1.Checked)
+                                // Is running
+                                foreach (var process in Process.GetProcessesByName("Terraria"))
                                 {
-                                    Console.WriteLine("The Terraria process was killed to continue operations.");
+                                    process.Kill();
+
+                                    // Log Item
+                                    if (checkBox1.Checked)
+                                    {
+                                        Console.WriteLine("The Terraria process was killed to continue operations.");
+                                    }
                                 }
                             }
                         }
 
-                        // Delete Folder
-                        Directory.Delete(Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text, true);
-
-                        // Log Item
-                        if (checkBox1.Checked)
+                        // Check if use steam directory.
+                        if (checkBox2.Checked)
                         {
-                            Console.WriteLine("Removed: " + Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text);
+                            // Get the parent directory.
+                            string OutDirParent = Directory.GetParent(Properties.Settings.Default.DepotPath).ToString();
+
+                            // Check if directory exists, if not, its the current version.
+                            if (Directory.Exists(OutDirParent + @"\Terraria-v" + itemRow.SubItems[0].Text))
+                            {
+                                // Exists, delete it.
+                                Directory.Delete(OutDirParent + @"\Terraria-v" + itemRow.SubItems[0].Text, true);
+
+                                // Log Item
+                                if (checkBox1.Checked)
+                                {
+                                    Console.WriteLine("Removed: " + OutDirParent + @"\Terraria-v" + itemRow.SubItems[0].Text);
+                                }
+                            }
+                            else
+                            {
+                                // Does not exist, log it.
+                                // Log Item
+                                if (checkBox1.Checked)
+                                {
+                                    Console.WriteLine("ERROR: Cannot delete Terraria-v" + itemRow.SubItems[0].Text + ". This version is currently active.");
+                                }
+
+                                // Display error.
+                                MessageBox.Show("ERROR: Cannot delete Terraria - v" + itemRow.SubItems[0].Text + ".\nThis version is currently active.", "ERROR: TerrariaDepotDownloader v" + FileVersionInfo.GetVersionInfo(Path.GetFileName(System.Windows.Forms.Application.ExecutablePath)).FileVersion, MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                            }
+                        }
+                        else
+                        {
+                            // Delete Folder
+                            Directory.Delete(Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text, true);
+
+                            // Log Item
+                            if (checkBox1.Checked)
+                            {
+                                Console.WriteLine("Removed: " + Properties.Settings.Default.DepotPath + @"\Terraria-v" + itemRow.SubItems[0].Text);
+                            }
                         }
 
                         // Update Forum
@@ -865,6 +890,9 @@ namespace TerrariaDepotDownloader
                     }
                 }
             }
+
+            // Edit Button
+            button5.Enabled = false;
         }
 
         // Update Button
